Restructured for less complexity.
